cardsorter/requirements.txt
python-pandas
python-numpy
tesseract-data-eng
